2023/23 - Switch graph representation
The improved graph representation allows part 1 to make use of the
graph built for part 2 by tracking whether each edge could be traversed
if we care what slopes are in the path.

Additionally, we stop tracking the exact points each edge traverses,
instead just keeping track of the path length.